refactor: extract HOOK_SETTINGS to shared module to prevent init/update drift
